backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
Pillow
python-multipart
ultralytics
opencv-python
# numpy # ultralyticsの依存関係でインストールされるはず
# tensorflow # Renderでのデプロイが困難なため、ここでは追加しません
# tensorflow-cpu